Move MLX scalar dispatches to scalar.py
