+++ pytensor/link/mlx/dispatch/__init__.py
@@ -3,6 +3,7 @@ from pytensor.link.mlx.dispatch.basic import mlx_funcify, mlx_typify
 
 import pytensor.link.mlx.dispatch.math
 import pytensor.link.mlx.dispatch.basic
+import pytensor.link.mlx.dispatch.scalar
 import pytensor.link.mlx.dispatch.elemwise
 import pytensor.link.mlx.dispatch.shape
 import pytensor.link.mlx.dispatch.subtensor
